query builder: simpler interface, better performance
Instead of storing and operating on a parsed AST, `Query` now stores the query
text as `[]byte`. We use `sqlp.Tokenizer` to parse inputs without generating an
AST, transcoding parameters on the fly. `IQuery` now simply appends to an
externally-passed `Query`, instead of having to return a parsed AST
representation. All together, this significantly simplifies the implementation
of `Query` and any external `IQuery` types.